refactor(history): extract ClipboardHistoryRow, drop unused @Bindable
